requirements.txt
nba_api>=1.2.1
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
psycopg[binary]>=3.1.10
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
pyarrow>=12.0.0
nfl-data-py
nflreadpy
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.24.0
db-dtypes>=1.2.0
pandas-gbq>=0.26.1
beautifulsoup4>=4.12.0
